% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReedFrost.R
\name{simulateReedFrost}
\alias{simulateReedFrost}
\title{A function to simulate epidemics using the Reed-Frost model}
\usage{
simulateReedFrost(Nsim = 1e+05, Npop = 10000, p = 0.00015, I0 = 1,
  Nit = 50, fig = T)
}
\description{
This function allows you to simulate epidemics.
}
\examples{
simulateReedFrost()
}
\keyword{Reed-Frost}
